Bounds are working (almost)
New file bounds.h where there is a map for all of the sciantix_variables, and some bounds corresponding (we need to complete that). For now error.txt is generated in the bin folder but at the end I hope it will be generated in the testFilePath.
We are calling the bounds function every time we change the variable. Maybe it is too heavy ; in fact this is too heavy because the time of execution is too slow doing that way