References done
Now the references appears in the documenation with a link to a pdf. You have to put the pdf in documentation/pdf_link. You have to put the name of the pdf file on the format : name_date. If it is not working, check the references.md file and modify the path in the tag <a>.