Bounds finished
From now on you can use bounds for : history variables, sciantix variables and system variables. To set its you have to go to setVariables for history and sciantix variables, and you have to use the setBounds and setBounded methods (already implemented, you jsut have to change the values).
For system_variables, you have to go to setSystem and you will have the same methods, at the end of every method like setBubbleDiffusivity.
In input settings there is a new option where you can : 0 - no check 1 - check with error and exit of the sciantix programm 2 - check with warnings with no exit.